pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orya-eval"
version = "0.1.0"
description = "Lightweight evaluation and regression checks for ML and AI systems."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Orya contributors" }
]
keywords = ["evaluation", "ml", "ai", "regression-testing", "metrics", "ci"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Testing",
]
dependencies = [
  "typer>=0.12,<1.0",
  "pydantic>=2.7,<3.0",
  "pandas>=2.2,<3.0",
  "PyYAML>=6.0,<7.0",
  "scikit-learn>=1.4,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1,<9.0",
  "ruff>=0.4,<1.0",
]

[project.scripts]
orya-eval = "orya_eval.cli:app"

[tool.setuptools.packages.find]
include = ["orya_eval*"]

[tool.setuptools.package-data]
orya_eval = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
